Removes Authorization initializer with parameters, documentation improvements
This is a breaking change

The removed Authorization initializer was effectively duplicative of the requestRights function, which in this framework is trivial to call right after creating it. Use of this initializer already wasn't one of the recommend usage patterns in the documentation.

Significantly improves the documentation for LaunchdManager.bless(...).